11/12/2025 11:40PM
I have added a Customer.java. In the Teller.java, the Bank.availableTellers.release() is not there anymore, instead it is in the Customer.java.
My thoughts: I have worked on the Customer.java. Originally, when I worked on the Customer.java, I realized I was going to encounter a deadlock. So, I added a assignedTeller.setBusy(false) to avoid a deadlock.
Plan: I plan to go over all the code, and see what needs to be fixed. I also plan to start testing with multiple customers. Eventually, I will make sure the program can handle 50 customers, and not just 3.